getorbuildExpandedDF now working for a single cell_type df - runs for SIM need to change all names to featureDF or expandedDF
